lockstep interpolation basically

+++ res/Shaders/default.vert
@@ -2,9 +2,9 @@
 
 ///////////////////////////////////////////////////////////
 
-layout (location = 0) in ivec2 inPos;
-layout (location = 1) in vec4  inCol;
-layout (location = 2) in uint  inTexId;
+layout (location = 0) in vec2 inPos;
+layout (location = 1) in vec4 inCol;
+layout (location = 2) in uint inTexId;
 
 ///////////////////////////////////////////////////////////
 
@@ -35,8 +35,8 @@ const vec2 TEXTURE_COORD [4] =
 
 void main() 
 {
-    float x = float(inPos.x) / meta.windowWidth  * 2 - 1;
-    float y = float(inPos.y) / meta.windowHeight * 2 - 1;
+    float x = inPos.x / meta.windowWidth  * 2 - 1;
+    float y = inPos.y / meta.windowHeight * 2 - 1;
     gl_Position = vec4(x, y, 0, 1);
 
     outCol = inCol;
